innit
WIP of the Voltage/Current Detection with Event Timer and Event Logging. Needs to work with the original Capstone_Yataki.ino file or potentially replace it. Will work on it later again when I feel like it. For now I need to work on the ACTUAL device itself, the device is still small in size, it needs to be bigger